finetune/LLMImage/requirements.txt
# Dependencias para fine-tuning de modelo de imagen a texto
transformers>=4.30.0
datasets>=2.12.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.5.0
mlflow>=2.5.0
accelerate>=0.20.0
jupyter>=1.0.0
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0 